Moved some code around for convenience
