+++ Core/GameObjects/Abstracts/Blueprint.cs
@@ -1,9 +1,9 @@
-﻿using Emberpoint.Core.GameObjects.Map;
+﻿using Emberpoint.Core.Extensions;
+using Emberpoint.Core.GameObjects.Map;
 using Microsoft.Xna.Framework;
 using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
-using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Runtime.CompilerServices;
@@ -91,8 +91,8 @@ namespace Emberpoint.Core.GameObjects.Abstracts
                     BlueprintTile tile = nullTile;
                     if (charValue !=null && !tiles.TryGetValue(charValue, out tile)) 
                         throw new Exception("Glyph '" + charValue + "' was not present in the config file for blueprint: " + name);
-                    var foregroundColor = GetColorByString(tile.Foreground);
-                    var backgroundColor = tile.Background != null ? GetColorByString(tile.Background) : Color.Black;
+                    var foregroundColor = MonoGameExtensions.GetColorByString(tile.Foreground);
+                    var backgroundColor = tile.Background != null ? MonoGameExtensions.GetColorByString(tile.Background) : Color.Black;
                     var cell = new T()
                     {
                         Glyph = tile.Glyph,
@@ -119,42 +119,13 @@ namespace Emberpoint.Core.GameObjects.Abstracts
 
                     if (!string.IsNullOrWhiteSpace(tile.LightColor))
                     {
-                        cell.LightProperties.LightColor = GetColorByString(tile.LightColor);
+                        cell.LightProperties.LightColor = MonoGameExtensions.GetColorByString(tile.LightColor);
                     }
 
                     cells.Add(cell);
                 }
             }
             return cells.ToArray();
-        }
-
-        private Color GetColorByString(string value)
-        {
-            if (value.StartsWith("#"))
-            {
-                value = value.TrimStart('#');
-                Color color;
-
-                if (value.Length == 6)
-                    color = new Color(
-                                int.Parse(value.Substring(0, 2), NumberStyles.HexNumber),
-                                int.Parse(value.Substring(2, 2), NumberStyles.HexNumber),
-                                int.Parse(value.Substring(4, 2), NumberStyles.HexNumber),
-                                255);
-                else // assuming length of 8
-                    color = new Color(
-                                int.Parse(value.Substring(2, 2), NumberStyles.HexNumber),
-                                int.Parse(value.Substring(4, 2), NumberStyles.HexNumber),
-                                int.Parse(value.Substring(6, 2), NumberStyles.HexNumber),
-                                int.Parse(value.Substring(0, 2), NumberStyles.HexNumber));
-
-                return color;
-            }
-
-            var prop = typeof(Color).GetProperty(value);
-            if (prop != null)
-                return (Color)prop.GetValue(null, null);
-            return default;
         }
     }
