+++ Core/UserInterface/Windows/FovWindow.cs
@@ -83,17 +83,10 @@ namespace Emberpoint.Core.UserInterface.Windows
             foreach (var character in characters)
             {
                 if (!_blueprintTiles.TryGetValue(character, out var tile) || tile.Name == null) continue;
-                var glyphColor = GetColorByString(tile.Foreground);
+                var glyphColor = MonoGameExtensions.GetColorByString(tile.Foreground);
+                if (glyphColor.A == 0) continue; // Don't render transparent tiles on the fov window
                 yield return new KeyValuePair<char, CharObj>(character, new CharObj(tile.Glyph, glyphColor, tile.Name));
             }
-        }
-
-        private Color GetColorByString(string value)
-        {
-            var prop = typeof(Color).GetProperty(value);
-            if (prop != null)
-                return (Color)prop.GetValue(null, null);
-            return default;
         }
 
         private void UpdateText()
